refactor: restructure backend by removing 'internal/' and 'cmd/yugi/' directories
