+++ backend/Dockerfile
@@ -6,7 +6,7 @@ WORKDIR /app
 COPY . .
 
 # Build the Go app
-RUN go build -o yugi ./cmd/yugi
+RUN go build -o yugi .
 
 # This container exposes port 8080 to the outside world
 EXPOSE 8080
