+++ backend/models/monster_card.go
